feature/auth-server client 연동 및 대응

+++ src/main/kotlin/com/swj9707/twittercloneapiserver/auth/controller/AuthController.kt
@@ -1,8 +1,9 @@
-package com.swj9707.twittercloneapiserver.Auth.controller
+package com.swj9707.twittercloneapiserver.auth.controller
 
-import com.swj9707.twittercloneapiserver.Auth.dto.UserReqDTO
-import com.swj9707.twittercloneapiserver.Auth.dto.UserResDTO
-import com.swj9707.twittercloneapiserver.Auth.service.TwitterUserService
+import com.swj9707.twittercloneapiserver.auth.dto.UserReqDTO
+import com.swj9707.twittercloneapiserver.auth.dto.UserResDTO
+import com.swj9707.twittercloneapiserver.auth.entity.TwitterUser
+import com.swj9707.twittercloneapiserver.auth.service.TwitterUserService
 import com.swj9707.twittercloneapiserver.constant.dto.BaseResponse
 import com.swj9707.twittercloneapiserver.constant.enum.BaseResponseCode
 import com.swj9707.twittercloneapiserver.exception.BaseException
@@ -11,12 +12,14 @@ import com.swj9707.twittercloneapiserver.utils.JwtUtil
 import jakarta.servlet.http.HttpServletRequest
 import jakarta.servlet.http.HttpServletResponse
 import org.springframework.http.ResponseEntity
+import org.springframework.security.core.annotation.AuthenticationPrincipal
 import org.springframework.web.bind.annotation.*
 
 @RestController
 @RequestMapping("/api/auth/v1")
 class AuthController(
     private val twitterUserService: TwitterUserService,
+    private val jwtUtil: JwtUtil,
     private val cookieUtil: CookieUtil) {
 
     @GetMapping("/test")
@@ -43,15 +46,22 @@ class AuthController(
         return ResponseEntity.ok().body(BaseResponse.success(result))
     }
 
+
     @PostMapping("/logout")
-    fun logout(@RequestBody userLogoutRes : UserReqDTO.Req.Logout,
-                @CookieValue(value = JwtUtil.REFRESH_TOKEN_NAME, defaultValue = "") refreshToken: String,
+    fun logout(@AuthenticationPrincipal user : TwitterUser,
+               @CookieValue(value = JwtUtil.REFRESH_TOKEN_NAME, defaultValue = "") refreshToken: String,
                req : HttpServletRequest, res : HttpServletResponse) : ResponseEntity<BaseResponse<UserResDTO.Res.Logout>> {
-        val result = twitterUserService.logout(userLogoutRes)
-        if(refreshToken.isNotEmpty()){
-            cookieUtil.deleteCookie(req, res, JwtUtil.REFRESH_TOKEN_NAME)
+
+        val accessToken = jwtUtil.resolveToken(req)
+        if(accessToken != null){
+            val result = twitterUserService.logout(accessToken)
+            if(refreshToken.isNotEmpty()){
+                cookieUtil.deleteCookie(req, res, JwtUtil.REFRESH_TOKEN_NAME)
+            }
+            return ResponseEntity.ok().body(BaseResponse.success(result))
+        } else {
+           throw BaseException(BaseResponseCode.INVALID_TOKEN)
         }
-        return ResponseEntity.ok().body(BaseResponse.success(result))
     }
 
     @PostMapping("/reissue")
